backend/es/requirements.txt
elasticsearch==8
python-dotenv
openai